trait separation and more tests
