+++ rust/personhog-replica/src/storage/traits/mod.rs
@@ -1,0 +1,11 @@
+mod cohort;
+mod distinct_id;
+mod feature_flag;
+mod group;
+mod person;
+
+pub use cohort::*;
+pub use distinct_id::*;
+pub use feature_flag::*;
+pub use group::*;
+pub use person::*;
