+++ rust/personhog-replica/src/storage/types/cohort.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Clone)]
+pub struct CohortMembership {
+    pub cohort_id: i64,
+    pub is_member: bool,
+}
