+++ rust/personhog-replica/src/storage/types/mod.rs
@@ -1,0 +1,9 @@
+mod cohort;
+mod feature_flag;
+mod group;
+mod person;
+
+pub use cohort::*;
+pub use feature_flag::*;
+pub use group::*;
+pub use person::*;
